+++ tests/end-to-end/backup-restore-test/Dockerfile
@@ -3,9 +3,9 @@ FROM golang:1.10-alpine as builder
 ENV SRC_DIR /go/src/github.com/kyma-project/kyma/tests/end-to-end/backup-restore-test
 WORKDIR ${SRC_DIR}
 
-COPY restore ${SRC_DIR}/restore/
+COPY . ${SRC_DIR}/
 
-RUN go test -c ./restore/ -o /restore.test
+RUN go test -c ./ -o /restore.test
 
 FROM alpine:3.8
 RUN apk --no-cache upgrade && apk --no-cache add curl
@@ -14,7 +14,6 @@ LABEL source = git@github.com:kyma-project/kyma.git
 RUN mkdir -p /root/.kube && touch /root/.kube/config
 
 COPY --from=builder /restore.test /restore.test
-COPY restore/function.yaml /function.yaml
 
 COPY entrypoint.sh /entrypoint.sh
