+++ tests/end-to-end/backup-restore-test/Gopkg.toml
@@ -26,6 +26,10 @@
 
 
 [[constraint]]
+  name = "github.com/google/uuid"
+  version = "1.1.0"
+
+[[constraint]]
   name = "github.com/smartystreets/goconvey"
   version = "1.6.3"
