+++ tests/end-to-end/backup-restore-test/backupe2e/function.go
@@ -13,7 +13,7 @@ import (
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/tools/clientcmd"
 
-	"github.com/kyma-project/kyma/tests/end-to-end/backup-restore-test/restore/utils"
+	"github.com/kyma-project/kyma/tests/end-to-end/backup-restore-test/utils"
 
 	. "github.com/smartystreets/goconvey/convey"
