+++ tests/end-to-end/backup-restore-test/before-commit.sh
@@ -47,7 +47,7 @@ fi
 ##
 # GO VET
 ##
-packagesToVet=("./restore/...")
+packagesToVet=("./...")
 
 for vPackage in "${packagesToVet[@]}"; do
 	vetResult=$(go vet ${vPackage})
